requirements-test.txt
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.6.1
requests-mock>=1.11.0 